[modify] move api.go to types/ directory
